requirements.txt
discord.py==2.7.1
python-dotenv==1.0.0